requirements.txt
streamlit
google-generativeai
chromadb
pypdf2
langchain
python-dotenv
pandas